requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Requests & APIs
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Google Sheets Integration
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
scrapy==2.11.0
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.25.2
thefuzz==0.20.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# CORS
fastapi-cors==0.0.6

# Logging
loguru==0.7.2

# Date/Time
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10